Update pipelines for AWS creds to assume a role  (#67)
* Update AWS configuration in CI/CD workflows

Modified the AWS configuration in both plugin-cd.yml and plugin-ci.yml to use a specific region (us-east-1) and assume a role for credentials, enabling role chaining. This change enhances security and aligns with best practices for AWS credential management.

* Refactor AWS region configuration in CI/CD workflows

Updated the AWS region configuration in both plugin-cd.yml and plugin-ci.yml to use a dynamic input variable instead of a hardcoded value. This change improves flexibility and allows for easier adjustments to the AWS region in future deployments.