remove chain role from actions (#72)
* Update AWS configuration in CI/CD workflows

Modified the AWS configuration in both plugin-cd.yml and plugin-ci.yml to use a specific region (us-east-1) and assume a role for credentials, enabling role chaining. This change enhances security and aligns with best practices for AWS credential management.

* Refactor AWS region configuration in CI/CD workflows

Updated the AWS region configuration in both plugin-cd.yml and plugin-ci.yml to use a dynamic input variable instead of a hardcoded value. This change improves flexibility and allows for easier adjustments to the AWS region in future deployments.

* hardcode aws region to us-east-1

* change var to secret for aws account id

* change to org secret for the aws account id of release

* enable write for id-token on permissions

* remove chain role from actions